--- src/main/java/racingcar/GameResultOutput.java
@@ -1,9 +1,0 @@
-package racingcar;
-
-import java.util.List;
-
-public interface GameResultOutput {
-    void printRoundResult(List<GameRoundResultOutput> results);
-
-    void printGameResult(List<String> winners);
-}
