--- src/main/java/racingcar/GameRoundResultOutput.java
@@ -1,5 +1,0 @@
-package racingcar;
-
-public interface GameRoundResultOutput {
-    String getGameRoundOutputFormat();
-}
