+++ src/main/java/racingcar/controller/RacingGameController.java
@@ -1,4 +1,13 @@
-package racingcar;
+package racingcar.controller;
+
+import racingcar.domain.car.Cars;
+import racingcar.domain.game.Game;
+import racingcar.domain.game.RacingGame;
+import racingcar.domain.game.RacingGameCount;
+import racingcar.view.input.GameInputView;
+import racingcar.view.input.RacingGameInputConsole;
+import racingcar.view.output.OutputView;
+import racingcar.view.output.RacingGameOutPutView;
 
 public class RacingGameController implements GameInputView {
     private final RacingGameInputConsole inputView;
@@ -20,7 +29,7 @@ public class RacingGameController implements GameInputView {
         String tryCountInput = inputView.readTryCountInput();
         RacingGameCount count = new RacingGameCount(tryCountInput);
 
-        GameResultOutput resultView = new RacingGameResultOutPut();
+        OutputView resultView = new RacingGameOutPutView();
         return new RacingGame(cars, count, resultView);
     }
 }
