+++ src/main/java/racingcar/domain/car/Car.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain.car;
 
 public class Car {
     private static final int MAX_NAME_LENGTH = 5;
