+++ src/main/java/racingcar/view/input/GameInputView.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.view.input;
 
 public interface GameInputView {
     void start();
