+++ src/main/java/racingcar/view/output/OutputView.java
@@ -1,0 +1,9 @@
+package racingcar.view.output;
+
+import java.util.List;
+
+public interface OutputView {
+    void printRoundResult(List<RoundView> results);
+
+    void printGameResult(List<String> winners);
+}
