+++ src/main/java/racingcar/view/output/RoundView.java
@@ -1,0 +1,5 @@
+package racingcar.view.output;
+
+public interface RoundView {
+    String getGameRoundOutputFormat();
+}
