+++ src/test/java/racingcar/CarTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.car.Car;
 
 @DisplayName("자동차 이름 검증 테스트")
 public class CarTest {
