+++ src/test/java/racingcar/CarsTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.car.Cars;
 
 @DisplayName("자동차들 테스트")
 public class CarsTest {
